refactor: db seed
testable 한 코드를 위한 리팩토링
bin/seed 에 있던 코드를 services/sequelize 로 옮기고,
그것을 실행하는 코드만 남김
